refactor: extract PerWorkerArenaLayout and arena-fit test helper
Centralise the per-worker Zone W byte walk in pippenger_arena_layout.hpp so
the sizer, live allocator, and arena-fit test all share one source of truth
(prior drift here caused four historical arena bugs). Move the 270-line
pippenger_bn254_arena_layout_fits_for_test helper out of
scalar_multiplication.cpp into the test TU and expose
compute_arena_bytes_for_msm via the header (BN254 explicitly instantiated).
